Add static_cast to fix size_t comparison warnings

+++ cfy_homework/src/task_3.cpp
@@ -55,13 +55,13 @@ Plan planning(const Target& target, const Eigen::Quaterniond& gimbal_quat) {
   static int stable_count = 0;               // 连续稳定帧计数
   
   omega_history.push_back(fitted_omega);
-  if (omega_history.size() > stable_frame_threshold) {
+  if (omega_history.size() > static_cast<size_t>(stable_frame_threshold)) {
     omega_history.pop_front();
   }
   
   // 检查历史窗口内的角速度是否稳定（变化率小于阈值）
   bool is_omega_stable = false;
-  if (omega_history.size() >= stable_frame_threshold) {
+  if (omega_history.size() >= static_cast<size_t>(stable_frame_threshold)) {
     float max_omega = *std::max_element(omega_history.begin(), omega_history.end());
     float min_omega = *std::min_element(omega_history.begin(), omega_history.end());
     float omega_range = max_omega - min_omega;
